Koitan saaha playerhealthia toimimaan

+++ Assets/Scripts/GlobalInt.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 [CreateAssetMenu(fileName = "GlobalInt", menuName = "ScriptableObjects/GlobalInt")]
-
 public class GlobalInt : ScriptableObject
 {
-    public float value;
+    public int value;
 }
